requirements.txt
reflex>=0.3.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
werkzeug>=3.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.5.0
